Avoid compiler warning. Added icons to configtoolboxdialog again.
[git-p4: depot-paths = "//depot/qt/main/": change = 74017]

+++ tools/designer/designer/listviewdnd.cpp
@@ -226,7 +226,7 @@ int ListViewDnd::buildTreeList( ListViewItemList &list )
 
 QListViewItem *ListViewDnd::itemAt( QPoint & pos )
 {
-    pos.ry() -= src->header()->height() * 1.5;
+    pos.ry() -= (int)(src->header()->height() * 1.5);
     QListViewItem *result = src->itemAt( pos );
     while ( result && result->parent() )
 	result = result->parent();
